+++ android/quest/src/main/res/layout/fragment_geo_widget_launcher.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.launcher.GeoWidgetLauncherFragment"
+    tools:context=".ui.geowidget.GeoWidgetLauncherFragment"
     >
 
     <androidx.fragment.app.FragmentContainerView
